// src/common/enums/permissions.enum.ts
export enum Permission {
  CREATE_ORGANIZATION = 'CREATE_ORGANIZATION',
  READ_ORGANIZATION = 'READ_ORGANIZATION',
  UPDATE_ORGANIZATION = 'UPDATE_ORGANIZATION',
  CREATE_USER = 'CREATE_USER',
  UPDATE_USER = 'UPDATE_USER',
  DELETE_USER = 'DELETE_USER',
  READ_USER = 'READ_USER',
  CREATE_BOT = 'CREATE_BOT',
  READ_BOT = 'READ_BOT',
  UPDATE_BOT = 'UPDATE_BOT',
  DELETE_BOT = 'DELETE_BOT',
  CREATE_KB = 'CREATE_KB',
  READ_KB = 'READ_KB',
  UPDATE_KB = 'UPDATE_KB',
  DELETE_KB = 'DELETE_KB',
  CREATE_KB_SOURCE = 'CREATE_KB_SOURCE',
  READ_KB_SOURCE = 'READ_KB_SOURCE',
  UPDATE_KB_SOURCE = 'UPDATE_KB_SOURCE',
  DELETE_KB_SOURCE = 'DELETE_KB_SOURCE',
  CREATE_COLLECTION = 'CREATE_COLLECTION',
  READ_COLLECTION = 'READ_COLLECTION',
  UPDATE_COLLECTION = 'UPDATE_COLLECTION',
  DELETE_COLLECTION = 'DELETE_COLLECTION',
  CREATE_TASK = 'CREATE_TASK',
  READ_TASK = 'READ_TASK',
  UPDATE_TASK = 'UPDATE_TASK',
  DELETE_TASK = 'DELETE_TASK',
  CREATE_WIDGET = 'CREATE_WIDGET',
  READ_WIDGET = 'READ_WIDGET',
  UPDATE_WIDGET = 'UPDATE_WIDGET',
  DELETE_WIDGET = 'DELETE_WIDGET',
  READ_AUDIT_LOG = 'READ_AUDIT_LOG',
  CREATE_PLAN = 'CREATE_PLAN',
  READ_PLAN = 'READ_PLAN',
  UPDATE_PLAN = 'UPDATE_PLAN',
  DELETE_PLAN = 'DELETE_PLAN',
  READ_SUBSCRIPTION = 'READ_SUBSCRIPTION',
  UPDATE_SUBSCRIPTION = 'UPDATE_SUBSCRIPTION',
  READ_TOKEN_WALLET = 'READ_TOKEN_WALLET',
  READ_TOKEN_TRANSACTION = 'READ_TOKEN_TRANSACTION',
  READ_PAYMENT_METHOD = 'READ_PAYMENT_METHOD',
  UPDATE_PAYMENT_METHOD = 'UPDATE_PAYMENT_METHOD',
  READ_INVOICE = 'READ_INVOICE',
}